test: split up two tests into two classes
